refactor: move to external package
