+++ service/federations/saml/groupmappings/doc.go
@@ -1,0 +1,3 @@
+// Package groupmappings provides a set of functions for interacting with
+// the Selectel SAML Federation Group Mappings API.
+package groupmappings
