+++ service/federations/saml/requests_test.go
@@ -15,10 +15,8 @@ import (
 )
 
 const (
-	federationsURL                    = "v1/federations/saml"
-	federationsIDURL                  = "v1/federations/saml/123"
-	federationGroupMappingsURL        = "v1/federations/saml/123/group-mappings"
-	federationExternalGroupMappingURL = "v1/federations/saml/123/group-mappings/456/external-groups/external-group"
+	federationsURL   = "v1/federations/saml"
+	federationsIDURL = "v1/federations/saml/123"
 )
 
 // Convenience vars for bool values.
@@ -535,340 +533,6 @@ func TestPreview(t *testing.T) {
 
 			require.ErrorIs(err, tt.expectedError)
 			assert.Equal(tt.expectedResponse, actual)
-		})
-	}
-}
-
-func TestGetGroupMappings(t *testing.T) {
-	tests := []struct {
-		name             string
-		prepare          func()
-		expectedResponse *GroupMappingsResponse
-		expectedError    error
-	}{
-		{
-			name: "ok",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodGet, testdata.TestURL+federationGroupMappingsURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusOK, testdata.TestGroupMappingsResponse)
-						return resp, nil
-					})
-			},
-			expectedResponse: &GroupMappingsResponse{
-				GroupMappings: []GroupMapping{
-					{
-						InternalGroupID: "456",
-						ExternalGroupID: "external-group",
-					},
-				},
-			},
-			expectedError: nil,
-		},
-		{
-			name: "error",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodGet, testdata.TestURL+federationGroupMappingsURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusForbidden, testdata.TestDoRequestErr)
-						return resp, nil
-					})
-			},
-			expectedResponse: nil,
-			expectedError:    iamerrors.ErrForbidden,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert := assert.New(t)
-			require := require.New(t)
-
-			federationsAPI := New(&client.BaseClient{
-				HTTPClient: &http.Client{},
-				APIUrl:     testdata.TestURL,
-				AuthMethod: &client.KeystoneTokenAuth{KeystoneToken: testdata.TestToken},
-			})
-
-			httpmock.ActivateNonDefault(federationsAPI.baseClient.HTTPClient)
-			defer httpmock.DeactivateAndReset()
-
-			tt.prepare()
-
-			ctx := context.Background()
-			actual, err := federationsAPI.GetGroupMappings(ctx, "123")
-
-			require.ErrorIs(err, tt.expectedError)
-			assert.Equal(tt.expectedResponse, actual)
-		})
-	}
-}
-
-func TestUpdateGroupMappings(t *testing.T) {
-	tests := []struct {
-		name             string
-		prepare          func()
-		input            GroupMappingsRequest
-		expectedResponse *GroupMappingsResponse
-		expectedError    error
-	}{
-		{
-			name: "ok",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodPatch, testdata.TestURL+federationGroupMappingsURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusOK, testdata.TestGroupMappingsResponse)
-						return resp, nil
-					})
-			},
-			input: GroupMappingsRequest{
-				GroupMappings: []GroupMapping{
-					{
-						InternalGroupID: "456",
-						ExternalGroupID: "external-group",
-					},
-				},
-			},
-			expectedResponse: &GroupMappingsResponse{
-				GroupMappings: []GroupMapping{
-					{
-						InternalGroupID: "456",
-						ExternalGroupID: "external-group",
-					},
-				},
-			},
-			expectedError: nil,
-		},
-		{
-			name: "error",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodPatch, testdata.TestURL+federationGroupMappingsURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusForbidden, testdata.TestDoRequestErr)
-						return resp, nil
-					})
-			},
-			input: GroupMappingsRequest{
-				GroupMappings: []GroupMapping{
-					{
-						InternalGroupID: "456",
-						ExternalGroupID: "external-group",
-					},
-				},
-			},
-			expectedResponse: nil,
-			expectedError:    iamerrors.ErrForbidden,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert := assert.New(t)
-			require := require.New(t)
-
-			federationsAPI := New(&client.BaseClient{
-				HTTPClient: &http.Client{},
-				APIUrl:     testdata.TestURL,
-				AuthMethod: &client.KeystoneTokenAuth{KeystoneToken: testdata.TestToken},
-			})
-
-			httpmock.ActivateNonDefault(federationsAPI.baseClient.HTTPClient)
-			defer httpmock.DeactivateAndReset()
-
-			tt.prepare()
-
-			ctx := context.Background()
-			actual, err := federationsAPI.UpdateGroupMappings(ctx, "123", tt.input)
-
-			require.ErrorIs(err, tt.expectedError)
-			assert.Equal(tt.expectedResponse, actual)
-		})
-	}
-}
-
-func TestAddExternalGroupMapping(t *testing.T) {
-	tests := []struct {
-		name          string
-		prepare       func()
-		expectedError error
-	}{
-		{
-			name: "ok",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodPut, testdata.TestURL+federationExternalGroupMappingURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusNoContent, "")
-						return resp, nil
-					})
-			},
-			expectedError: nil,
-		},
-		{
-			name: "error",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodPut, testdata.TestURL+federationExternalGroupMappingURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusForbidden, testdata.TestDoRequestErr)
-						return resp, nil
-					})
-			},
-			expectedError: iamerrors.ErrForbidden,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			require := require.New(t)
-
-			federationsAPI := New(&client.BaseClient{
-				HTTPClient: &http.Client{},
-				APIUrl:     testdata.TestURL,
-				AuthMethod: &client.KeystoneTokenAuth{KeystoneToken: testdata.TestToken},
-			})
-
-			httpmock.ActivateNonDefault(federationsAPI.baseClient.HTTPClient)
-			defer httpmock.DeactivateAndReset()
-
-			tt.prepare()
-
-			ctx := context.Background()
-			err := federationsAPI.AddExternalGroupMapping(ctx, "123", "456", "external-group")
-
-			require.ErrorIs(err, tt.expectedError)
-		})
-	}
-}
-
-func TestDeleteExternalGroupMapping(t *testing.T) {
-	tests := []struct {
-		name          string
-		prepare       func()
-		expectedError error
-	}{
-		{
-			name: "ok",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodDelete, testdata.TestURL+federationExternalGroupMappingURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusNoContent, "")
-						return resp, nil
-					})
-			},
-			expectedError: nil,
-		},
-		{
-			name: "error",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodDelete, testdata.TestURL+federationExternalGroupMappingURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusForbidden, testdata.TestDoRequestErr)
-						return resp, nil
-					})
-			},
-			expectedError: iamerrors.ErrForbidden,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			require := require.New(t)
-
-			federationsAPI := New(&client.BaseClient{
-				HTTPClient: &http.Client{},
-				APIUrl:     testdata.TestURL,
-				AuthMethod: &client.KeystoneTokenAuth{KeystoneToken: testdata.TestToken},
-			})
-
-			httpmock.ActivateNonDefault(federationsAPI.baseClient.HTTPClient)
-			defer httpmock.DeactivateAndReset()
-
-			tt.prepare()
-
-			ctx := context.Background()
-			err := federationsAPI.DeleteExternalGroupMapping(ctx, "123", "456", "external-group")
-
-			require.ErrorIs(err, tt.expectedError)
-		})
-	}
-}
-
-func TestExternalGroupMappingExists(t *testing.T) {
-	tests := []struct {
-		name           string
-		prepare        func()
-		expectedExists bool
-		expectedError  error
-	}{
-		{
-			name: "exists",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodHead, testdata.TestURL+federationExternalGroupMappingURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusNoContent, "")
-						return resp, nil
-					})
-			},
-			expectedExists: true,
-			expectedError:  nil,
-		},
-		{
-			name: "not found",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodHead, testdata.TestURL+federationExternalGroupMappingURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusNotFound, testdata.TestUserOrGroupNotFoundErr)
-						return resp, nil
-					})
-			},
-			expectedExists: false,
-			expectedError:  nil,
-		},
-		{
-			name: "error",
-			prepare: func() {
-				httpmock.RegisterResponder(
-					http.MethodHead, testdata.TestURL+federationExternalGroupMappingURL,
-					func(r *http.Request) (*http.Response, error) {
-						resp := httpmock.NewStringResponse(http.StatusForbidden, testdata.TestDoRequestErr)
-						return resp, nil
-					})
-			},
-			expectedExists: false,
-			expectedError:  iamerrors.ErrForbidden,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert := assert.New(t)
-			require := require.New(t)
-
-			federationsAPI := New(&client.BaseClient{
-				HTTPClient: &http.Client{},
-				APIUrl:     testdata.TestURL,
-				AuthMethod: &client.KeystoneTokenAuth{KeystoneToken: testdata.TestToken},
-			})
-
-			httpmock.ActivateNonDefault(federationsAPI.baseClient.HTTPClient)
-			defer httpmock.DeactivateAndReset()
-
-			tt.prepare()
-
-			ctx := context.Background()
-			exists, err := federationsAPI.ExternalGroupMappingExists(ctx, "123", "456", "external-group")
-
-			require.ErrorIs(err, tt.expectedError)
-			assert.Equal(tt.expectedExists, exists)
 		})
 	}
 }
